rename files to kebab-case
