+++ backend/src/middlewares/authenticate-user.ts
