+++ frontend/src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
-import { ThemeProvider } from "../hooks/theme";
-import Navbar from "./Navbar";
+import { ThemeProvider } from "@hooks/use-theme";
+import Navbar from "@components/navbar";
 
 const Layout = () => {
   return (
