+++ frontend/src/components/dashboard/chat/comment-item.tsx
