+++ frontend/src/components/dashboard/chat/message-input.tsx
