+++ frontend/src/components/dashboard/discussion-info-panel.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { useAuth } from "../../hooks/auth";
+import { useAuth } from "../../hooks/use-auth";
 import { Card, CardContent, CardHeader, CardTitle } from "@components/ui/card";
 import { Button } from "@components/ui/button";
 import { Textarea } from "@components/ui/textarea";
