+++ frontend/src/components/dashboard/discussions-sidebar.tsx
