+++ frontend/src/components/dashboard/modals/delete-discussion-modal.tsx
@@ -5,8 +5,8 @@ import {
   DialogFooter,
   DialogHeader,
   DialogTitle,
-} from "../../ui/dialog";
-import { Button } from "../../ui/button";
+} from "@components/ui/dialog";
+import { Button } from "@components/ui/button";
 
 interface DeleteDiscussionModalProps {
   isOpen: boolean;
