+++ frontend/src/components/dashboard/modals/delete-group-modal.tsx
