+++ frontend/src/components/dashboard/modals/edit-discussion-modal.tsx
@@ -7,12 +7,12 @@ import {
   DialogFooter,
   DialogHeader,
   DialogTitle,
-} from "../../ui/dialog";
-import { Button } from "../../ui/button";
-import { Input } from "../../ui/input";
-import { Label } from "../../ui/label";
-import { Textarea } from "../../ui/textarea";
-import { Discussion } from "../../../lib/api/discussion";
+} from "@components/ui/dialog";
+import { Button } from "@components/ui/button";
+import { Input } from "@components/ui/input";
+import { Label } from "@components/ui/label";
+import { Textarea } from "@components/ui/textarea";
+import { Discussion } from "@lib/api/discussion";
 
 interface EditDiscussionModalProps {
   discussion: Discussion | null;
