+++ frontend/src/components/dashboard/modals/join-group-modal.tsx
