+++ frontend/src/components/protected-route.tsx
@@ -1,6 +1,6 @@
 import { ReactNode } from "react";
 import { Navigate } from "react-router-dom";
-import { useAuth } from "../hooks/auth";
+import { useAuth } from "@hooks/use-auth";
 
 type ProtectedRouteProps = {
   children: ReactNode;
