+++ frontend/src/components/ui/mode-toggle.tsx
@@ -7,7 +7,7 @@ import {
   DropdownMenuItem,
   DropdownMenuTrigger,
 } from "@components/ui/dropdown-menu";
-import { useTheme } from "@hooks/theme";
+import { useTheme } from "@hooks/use-theme";
 
 export function ModeToggle() {
   const { setTheme } = useTheme();
