+++ frontend/src/hooks/use-chat.ts
